# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateStreamRequest:
    properties:
      category:
        example: Gaming
        maxLength: 50
        type: string
      description:
        example: This is my streaming description
        maxLength: 500
        type: string
      is_mature:
        example: false
        type: boolean
      is_public:
        example: true
        type: boolean
      language:
        example: en
        maxLength: 10
        type: string
      tags:
        example:
        - gaming
        - live
        - fun
        items:
          type: string
        maxItems: 10
        type: array
      title:
        example: My Awesome Stream
        maxLength: 100
        type: string
    required:
    - title
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      confirm_password:
        example: password123
        type: string
      display_name:
        example: John Doe
        maxLength: 50
        type: string
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: johndoe
        maxLength: 20
        minLength: 3
        type: string
    required:
    - confirm_password
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Live streaming platform API with chat
  title: GoLive API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Account deactivated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logout user and clear authentication cookie
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Auth
  /auth/me:
    get:
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Refresh JWT token using existing token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Account deactivated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh JWT Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email or username already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User Registration
      tags:
      - Auth
...
  /me/streams:
    get:
      description: Get list of streams for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User's streams
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User's Streams
      tags:
      - Streams
    post:
      consumes:
      - application/json
      description: Create a new stream for the authenticated user
      parameters:
      - description: Stream creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stream created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already has an active stream
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create New Stream
      tags:
      - Streams
  /me/streams/{key}:
    delete:
      description: Delete a stream (owner only)
      parameters:
      - description: Stream Key
        in: path
        name: key
        required: true
        type: string
      responses:
        "200":
          description: Stream deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Stream key required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - not stream owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Stream not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Stream
      tags:
      - Streams
    put:
      consumes:
      - application/json
      description: Update stream information (owner only)
      parameters:
      - description: Stream Key
        in: path
        name: key
        required: true
        type: string
      - description: Stream update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stream updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error or no updates
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - not stream owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Stream not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Stream
      tags:
      - Streams
  /me/streams/{key}/start:
    post:
      description: Start a stream (set it as live)
      parameters:
      - description: Stream Key
        in: path
        name: key
        required: true
        type: string
      responses:
        "200":
          description: Stream started successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Stream key required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Start Stream
      tags:
      - Streams
  /me/streams/{key}/stop:
    post:
      description: Stop a stream (set it as offline)
      parameters:
      - description: Stream Key
        in: path
        name: key
        required: true
        type: string
      responses:
        "200":
          description: Stream stopped successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Stream key required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Stop Stream
      tags:
      - Streams
...
  /realtime/channels:
    get:
      description: Retourne la liste complète des channels en mémoire.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChannelListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lister les channels actifs
      tags:
      - Realtime
  /realtime/channels/{channel}:
    get:
      description: Retourne la configuration et les statistiques instantanées du channel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realtime.ChannelInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Récupérer les informations d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/access:
    post:
      description: Ajoute une liste d'utilisateurs autorisés sur un channel privé.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Liste des utilisateurs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChannelAccessRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Autoriser l'accès à un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/access/{userID}:
    delete:
      description: Supprime l'accès d'un utilisateur à un channel privé.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: ID utilisateur
        in: path
        name: userID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Révoquer l'accès d'un utilisateur
      tags:
      - Realtime
  /realtime/channels/{channel}/broadcast:
    post:
      description: Envoie un message système à tous les clients abonnés.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Payload système
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BroadcastRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BroadcastResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Diffuser un message système sur un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/clear:
    post:
      description: Efface l'historique d'un channel et avertit les clients.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Vider un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/history:
    get:
      description: Retourne les derniers messages d'un channel temps réel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Nombre maximum de messages
        in: query
        maximum: 1000
        name: limit
        type: integer
      - description: Identifiant du message à partir duquel récupérer
        in: query
        name: since
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Récupérer l'historique d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/messages/{messageID}:
    delete:
      description: Supprime un message spécifique et informe les clients connectés.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Identifiant du message
        in: path
        name: messageID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Supprimer un message d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/presence:
    get:
      description: Retourne la liste des utilisateurs actuellement connectés sur le
        channel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PresenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Récupérer la présence d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/presence/sync:
    post:
      description: Remplace la présence en mémoire par la liste fournie (usage déporté).
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Liste de présence
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PresenceSyncRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Synchroniser la présence d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/publish:
    post:
      description: Publie un message pour tous les clients abonnés au channel spécifié.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Payload du message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PublishRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PublishResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publier un message sur un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/settings:
    put:
      description: Ajuste les limites et règles d'accès d'un channel temps réel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: Paramètres du channel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChannelSettingsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realtime.ChannelSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mettre à jour les paramètres d'un channel
      tags:
      - Realtime
  /realtime/channels/{channel}/users/{userID}/ban:
    delete:
      description: Retire un ban précédemment appliqué et notifie le channel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: ID utilisateur ciblé
        in: path
        name: userID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Débannir un utilisateur
      tags:
      - Realtime
    post:
      description: Empêche définitivement un utilisateur de publier sur le channel.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: ID utilisateur ciblé
        in: path
        name: userID
        required: true
        type: string
      - description: Motif du ban
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BanRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BanResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bannir un utilisateur
      tags:
      - Realtime
  /realtime/channels/{channel}/users/{userID}/timeout:
    post:
      description: Applique un timeout temporaire sur un utilisateur pour le channel
        cible.
      parameters:
      - description: Channel ID (namespace:identifiant)
        in: path
        name: channel
        required: true
        type: string
      - description: ID utilisateur ciblé
        in: path
        name: userID
        required: true
        type: string
      - description: Paramètres du timeout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TimeoutRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TimeoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Timeout un utilisateur
      tags:
      - Realtime
  /realtime/rpc:
    post:
      description: Exécute une méthode RPC exposée par le serveur realtime.
      parameters:
      - description: Requête RPC
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RPCRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RPCResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Appeler un RPC realtime
      tags:
      - Realtime
  /realtime/stats:
    get:
      description: Retourne des informations agrégées sur les channels et clients
        connectés.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realtime.ServerStats'
      security:
      - BearerAuth: []
      summary: Obtenir les statistiques du serveur realtime
      tags:
      - Realtime
  /streams:
    get:
      description: Get list of public streams with pagination and filtering
      parameters:
      - default: 20
        description: Number of streams to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of streams to skip
        in: query
        name: offset
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter only live streams
        in: query
        name: live
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of streams with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List Public Streams
      tags:
      - Streams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"